Last hit detection and Death/Game Reset
Ball detects which player hit the ball last by player ID and ignores hurtbox of that player. Changes colour to indicate who hit it last.
Basic death/game end and reset disables controls and reloads scene after 3 seconds. Forgot to put dontdestroyonload for stuff so you have to rejoin the game each time but shouldn't need to be implemented if this just becomes a minigame.